test branch for fixing android builds
Contains part of the official gallery example from the Qt package.
Idea is to build the same thing with QtCreator and qmake especially,
and compare the result with a build via cmake and our ecosystem here.

The branch should never be merged into master. The other way round is
ok and useful of course.